% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeData.R
\name{changeData}
\alias{changeData}
\title{Creates a list containing the raster datasets concerning the number
of times the category of interest is present and the number of times
the category of interest changes during the time series.}
\usage{
changeData(x, nodata = 2, category = 1, spres = c(1000, 1000), datacrs = NULL)
}
\arguments{
\item{x}{is the data, which must be a RasterStack,RasterBrick,SpatRaster or
data.frame.}

\item{nodata}{is alpahnumeric, which denotes no data in the dataset.}

\item{category}{is the category of interest. The default is set to 1.}

\item{spres}{is a 1*2 vector indicating the spatial resolution of the data.
The default is set to c(1000,1000)}

\item{datacrs}{is the Coordinate Reference System (CRS) of the input data.}
}
\value{
The output from \code{\link{changeData}}
}
\description{
Creates a list containing the raster datasets concerning the number
of times the category of interest is present and the number of times
the category of interest changes during the time series.
}
\examples{
\dontrun{changeData(x = data,nodata = 2,category = 1,spres = c(1000,1000),
datacrs = "+proj=utm +zone=32 +datum=WGS84 +ellps=GRS80  +units=m +no_defs")
}
}
