% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStack.R
\name{dataStack}
\alias{dataStack}
\title{Creates results that serve as input for the stackbarPlot function.}
\usage{
dataStack(
  x,
  category = 1,
  nodata = 2,
  unified = "yes",
  timePoints = c(2000, 2001, 2002, 2003, 2005),
  categoryName = "marsh",
  regionName = "Study Region"
)
}
\arguments{
\item{x}{is the data, which must be a RasterStack,RasterBrick,SpatRaster or data.frame.}

\item{category}{is the category of interest. The default is set to 1.}

\item{nodata}{is alphanumeric, which denotes no data in the dataset.}

\item{unified}{is a string, which can be "yes" or "no" only. If "yes," the change is a percentage of a region's
unified area; else, the change is a percentage of the  entire region under consideration.}

\item{timePoints}{is a vector containg the time points under consideration.The default is  c(2000,2001,2002,2003,2005).}

\item{categoryName}{is a character representing the name of the category of interest.Default is "category"}

\item{regionName}{is a string or character the name of the study region.}
}
\value{
The output from \code{\link{dataStack}}
}
\description{
Creates results that serve as input for the stackbarPlot function.
}
\examples{
\dontrun{trajPlot(x,
}
}
