% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackbarPlots.R
\name{stackbarPlots}
\alias{stackbarPlots}
\title{Create stack bar plots showing trajectories' sizes and the three change components.}
\usage{
stackbarPlots(
  input,
  axisSize = 12,
  lbAxSize = 15,
  lgSize = 12,
  titleSize = 15,
  datbreaks = "no",
  upperlym = 35,
  lowerlym = -50,
  lymby = 5,
  upperlym2 = 0.5,
  lymby2 = 0.1,
  xAngle = 0
)
}
\arguments{
\item{input}{is the results from running the dataStack function.}

\item{axisSize}{is a numerical value that control the size of the labels on tick marks of the horizontal
and vertical tick marks.}

\item{lbAxSize}{is a numerical value to control the size of the labels on the horizontal and verictcal axis.}

\item{lgSize}{is a numerical value to control the size of the legend text.}

\item{titleSize}{is a numerical value to control the size of the title text.}

\item{datbreaks}{is a string of "yes" or "no", which controls the range and sub-division of the vertical axis
of the stacked bar plots.}

\item{upperlym}{if datbreaks set to "yes," is a numerical value to control the upper limit of the trajecoty stackbar plot.}

\item{lowerlym}{if datbreaks set to "yes," is a numerical value to control the lower limit of the trajecoty stackbar plot.}

\item{lymby}{if datbreaks set to "yes," is a numerical value to control interval on the vertical axis of the components of change stackbar plot.}

\item{upperlym2}{if datbreaks set to "yes," is a numerical value to control the upper limit of the components of change stacked bar plot.}

\item{lymby2}{if datbreaks set to "yes," is a numerical value to control the interval on the vertical axis of the components of change stacked bar plot.}

\item{xAngle}{is a numerical value to control the orientation of the text on the vertical axis of the trajecoty stackbar plot}
}
\value{
The output from \code{\link{stackbarPlots}}
}
\description{
Create stack bar plots showing trajectories' sizes and the three change components.
}
\examples{
\dontrun{stackbarPlots(input,
                       axisSize = 12,
                       lbAxSize = 15,
                       lgSize = 12,
                       titleSize = 15,
                       upperlym = 35,
                       lowerlym = - 50,
                       lymby = 5,
                       upperlym2 = 0.5,
                       lymby2 = 0.1,
                       xAngle = 0)
}
}
